# Backend/src/Aurigma.DirectMail.Sample.TypeScriptApiClient/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Direct Mail API",
    "version": "v1"
  },
  "paths": {
...
    "/api/direct-mail/v1/jobs": {
      "get": {
        "tags": [
          "Job"
        ],
        "summary": "Returns a list of jobs.",
        "operationId": "Job_GetJobs",
        "parameters": [
          {
            "name": "LineItemId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/direct-mail/v1/jobs/{id}/restart": {
      "post": {
        "tags": [
          "Job"
        ],
        "summary": "Restarts a job.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/direct-mail/v1/jobs/{id}/download-results": {
      "post": {
        "tags": [
          "Job"
        ],
        "summary": "Returns a collection of processing results.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobProcessingResultDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobProcessingResultDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobProcessingResultDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConflictDto": {
        "type": "object",
        "properties": {
          "propertyName": {
            "type": "string",
            "nullable": true
          },
          "propertyValue": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "JobDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Job id.",
            "format": "uuid"
          },
          "customersCanvasProjectId": {
            "type": "integer",
            "description": "Customer's Canvas project id.",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "additionalProperties": false
      },
      "JobProcessingResultDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "File url.",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "File name.",
            "nullable": true
          },
          "format": {
            "type": "string",
            "description": "File format.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "JobStatus": {
        "enum": [
          "Pending",
          "InProgress",
          "Completed",
          "Failed"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}